Migration stable

+++ database/migrations/2023_08_09_212004_create_objectives_table.php
@@ -28,7 +28,7 @@ return new class extends Migration
             $table->longText('description')->nullable();
 
             $table->dateTime('deadline')->nullable();
-            $table->float('goal')->nullable();
+            $table->decimal('goal', 8,2)->nullable();
 
             $table->boolean('draft')->default(1); // it's a draft on the assignment and can be adjusted in another view.
             $table->softDeletes();
